backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Redis & Background Jobs
redis==5.0.1
celery==5.3.6

# PDF & Excel Generation
reportlab==4.0.9
weasyprint==60.2
openpyxl==3.1.2
xlsxwriter==3.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# CORS & Security
python-cors==1.0.0
cryptography==42.0.1

# AWS S3 (Optional - for future upgrade)
boto3==1.34.34
